Set up config for aliases and basic css vars
