+++ craco.config.js
@@ -1,0 +1,10 @@
+const path = require('path');
+
+module.exports = {
+    webpack: {
+        alias: {
+            assets: path.resolve(__dirname, 'src/assets/'),
+            components: path.resolve(__dirname, 'src/components/'),
+        },
+    },
+};
