+++ src/components/Footer/Footer.module.css
@@ -1,0 +1,9 @@
+.footer {
+  width: 100%;
+}
+
+.img {
+  object-fit: cover;
+  width: 100%;
+  height: auto;
+}
